feat: add Handle strict parser conforming to LexiconStringFormat
